% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp2domains_by_twtest.R
\name{comp2domains_by_twtest}
\alias{comp2domains_by_twtest}
\title{compare single pair of domains}
\usage{
comp2domains_by_twtest(
  chrn,
  start,
  end,
  fhic0,
  fhic1,
  ofile = paste(getwd(), "/stdout.txt", sep = ""),
  min_nbin = 10,
  hicnorm = "KR",
  reso = 100000,
  f = 0.5
)
}
\arguments{
\item{chrn}{chromosomee to compare}

\item{start}{start bin of chrn}

\item{end}{end bin of chrn}

\item{fhic0}{the first hic file to be compared}

\item{fhic1}{the  second hic file to be compared}

\item{ofile}{the file path of the result}

\item{min_nbin}{default is 10}

\item{hicnorm}{default is "KR"}

\item{reso}{resolution to choose,default is 100000}

\item{f}{default is 0.5}
}
\value{
result of comparition
}
\description{
compare single pair of domains
}
\examples{
comp2domains_by_twtest('/lanec1_home/data/GSE63525_cell2014/GSE63525_GM12878_insitu_primary_replicate_combined.hic',
'/lanec1_home/data/GSE63525_cell2014/GSE63525_K562_combined.hic',
ofile = '/lanec2_home/zhangx/DiffCompare/pythonR_compare_result/stdout_R')
}
