% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nbins.R
\name{compute_nbins}
\alias{compute_nbins}
\title{return how many bins between start and end, based on the chosen resolution}
\usage{
compute_nbins(start, end, reso)
}
\arguments{
\item{start}{start bin of chromosome}

\item{end}{end bin of chromosome}

\item{reso}{chosen resolution}
}
\value{
number of bins
}
\description{
return how many bins between start and end, based on the chosen resolution
}
\examples{
compute_nbins(1060000,1080000,10000)
}
