% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractKdiagonalCsrMatrix.R
\name{extractKdiagonalCsrMatrix}
\alias{extractKdiagonalCsrMatrix}
\title{make matrix norm}
\usage{
extractKdiagonalCsrMatrix(spsCsrMat)
}
\arguments{
\item{spsCsrMat}{matrix}
}
\value{
matrix
}
\description{
make matrix norm
}
\examples{
extractKdiagonalCsrMatrix(matrix)
}
