% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonZero.R
\name{NonZero}
\alias{NonZero}
\title{Find those whose value equal zero}
\usage{
NonZero(D)
}
\arguments{
\item{D}{matrix}
}
\value{
whose value equals zero and their rows and cols
}
\description{
Find those whose value equal zero
}
\examples{
NonZero(input)
}
